requirements.txt
Django==2.0.7
django-filter==2.0.0
django-filters==0.2.1
django-model-utils==3.1.2
jupyter==1.0.0
Pillow==5.2.0
psycopg2==2.7.5
pytz==2018.5
